Move Utils to cpp
